Changes on restructring of the NTAP MODEL files.
